Rework extension
The old hook used no longer functions, rework extension to add a
variable called {{FILECATEGORIES}} instead which adds the page to the
requisite categories. It is automatically added on new uploads, but must
be manually added to existing ones.

Additionally, since the ID3 parsing embedded in the extension was a bit
spotty, use an ID3 library instead. The library chosen is available on
composer, seems actively maintained. The library is available under
multiple open source licenses, GPLv3 was chosen since this extension is
GPLv3 as well :)

Change-Id: Ia5c954c423d38f8b973d0bcbd226e07d10dc8800